Error investigation
both quaternion propagation and acceleration are the sources of error.

After feeding true quaternions into the v and r propagator the error persisted. So the problem is on the acceleration computation, as v and r are not always different but instead diverge as time progresses.

The error in quaternions remains. Analyzing quaternion propagation it is found that the quaternions slowly diverge. A possible source of errors might be that the quaternion propagation is correct but the alpha, beta, sigma computation in the simulator is incorrect. THE ATTITUDE EQS IN THE SIMULATOR WHERE SIMPLIFIED AND DIDN'T ACCOUNT FOR CHANGING LONGITUDE AND LATITUDE (circular planet).